♻️ Move slog.go to public and add aliases
